Move some types to common package
