+++ packages/extension/src/background/messaging/internalMethods.ts
@@ -2,10 +2,11 @@ import algosdk from 'algosdk';
 import { JsonRpcMethod } from '@algosigner/common/messaging/types';
 import { logging } from '@algosigner/common/logging';
 import { ExtensionStorage } from '@algosigner/storage/src/extensionStorage';
+import { Ledger, Namespace } from '@algosigner/common/types';
+import { AliasConfig } from '@algosigner/common/config';
 import { Task } from './task';
-import { API, Cache, Ledger } from './types';
-import { Namespace } from '@algosigner/common/types';
-import { Settings, AliasConfig } from '../config';
+import { API, Cache } from './types';
+import { Settings,  } from '../config';
 import encryptionWrap from '../encryptionWrap';
 import Session from '../utils/session';
 import AssetsDetailsHelper from '../utils/assetsDetailsHelper';
