+++ packages/extension/src/background/messaging/types.ts
@@ -5,10 +5,6 @@ import { LedgerTemplate } from '@algosigner/common/types/ledgers';
 // Key and value must match in this enum so we
 // can compare its existance with i.e. "TestNet" in SupportedLedger
 /* eslint-disable no-unused-vars */
-export enum Ledger {
-  TestNet = 'TestNet',
-  MainNet = 'MainNet',
-}
 
 export enum Backend {
   PureStake = 'PureStake',
